objects: reorganise dir structure
