+++ meson.build
@@ -132,11 +132,7 @@ sources = [
   'src/game/los.c',
   'src/game/lot.c',
   'src/game/music.c',
-  'src/game/objects.c',
-  'src/game/objects/boat.c',
-  'src/game/objects/bridge.c',
-  'src/game/objects/cabin.c',
-  'src/game/objects/cog.c',
+  'src/game/objects/common.c',
   'src/game/objects/creatures/ape.c',
   'src/game/objects/creatures/bacon_lara.c',
   'src/game/objects/creatures/baldy.c',
@@ -160,7 +156,6 @@ sources = [
   'src/game/objects/creatures/torso.c',
   'src/game/objects/creatures/trex.c',
   'src/game/objects/creatures/wolf.c',
-  'src/game/objects/door.c',
   'src/game/objects/effects/blood.c',
   'src/game/objects/effects/body_part.c',
   'src/game/objects/effects/bubble.c',
@@ -172,14 +167,19 @@ sources = [
   'src/game/objects/effects/splash.c',
   'src/game/objects/effects/twinkle.c',
   'src/game/objects/effects/waterfall.c',
-  'src/game/objects/keyhole.c',
-  'src/game/objects/misc.c',
-  'src/game/objects/pickup.c',
-  'src/game/objects/puzzle_hole.c',
-  'src/game/objects/save_crystal.c',
-  'src/game/objects/scion.c',
-  'src/game/objects/switch.c',
-  'src/game/objects/trapdoor.c',
+  'src/game/objects/general/boat.c',
+  'src/game/objects/general/bridge.c',
+  'src/game/objects/general/cabin.c',
+  'src/game/objects/general/cog.c',
+  'src/game/objects/general/door.c',
+  'src/game/objects/general/keyhole.c',
+  'src/game/objects/general/misc.c',
+  'src/game/objects/general/pickup.c',
+  'src/game/objects/general/puzzle_hole.c',
+  'src/game/objects/general/save_crystal.c',
+  'src/game/objects/general/scion.c',
+  'src/game/objects/general/switch.c',
+  'src/game/objects/general/trapdoor.c',
   'src/game/objects/traps/damocles_sword.c',
   'src/game/objects/traps/dart.c',
   'src/game/objects/traps/falling_block.c',
