+++ src/game/collide.c
@@ -3,8 +3,7 @@
 #include "config.h"
 #include "game/items.h"
 #include "game/lara.h"
-#include "game/objects.h"
-#include "game/objects/door.h"
+#include "game/objects/general/door.h"
 #include "game/room.h"
 #include "game/sound.h"
 #include "global/const.h"
