+++ src/game/objects/common.h
