+++ src/game/objects/creatures/pod.c
@@ -4,7 +4,7 @@
 #include "game/gamebuf.h"
 #include "game/items.h"
 #include "game/lot.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "game/sound.h"
 #include "global/vars.h"
