+++ src/game/objects/creatures/skate_kid.c
@@ -3,7 +3,7 @@
 #include "game/creature.h"
 #include "game/items.h"
 #include "game/music.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "game/random.h"
 #include "global/const.h"
 #include "global/vars.h"
