+++ src/game/objects/effects/earthquake.c
@@ -1,7 +1,7 @@
 #include "game/objects/effects/earthquake.h"
 
 #include "game/items.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "game/random.h"
 #include "game/sound.h"
 #include "global/vars.h"
