+++ src/game/objects/general/boat.c
@@ -1,4 +1,4 @@
-#include "game/objects/boat.h"
+#include "game/objects/general/boat.h"
 
 #include "game/items.h"
 #include "global/vars.h"
