+++ src/game/objects/general/cabin.h
