+++ src/game/objects/general/cog.h
