+++ src/game/objects/general/keyhole.h
