+++ src/game/objects/general/misc.h
