+++ src/game/objects/general/puzzle_hole.h
