+++ src/game/objects/general/save_crystal.h
