+++ src/game/objects/general/scion.c
@@ -1,12 +1,12 @@
-#include "game/objects/scion.h"
+#include "game/objects/general/scion.h"
 
 #include "game/effects.h"
 #include "game/input.h"
 #include "game/inv.h"
 #include "game/items.h"
 #include "game/lara.h"
-#include "game/objects.h"
-#include "game/objects/pickup.h"
+#include "game/objects/common.h"
+#include "game/objects/general/pickup.h"
 #include "game/overlay.h"
 #include "game/random.h"
 #include "game/room.h"
