+++ src/game/objects/general/scion.h
