+++ src/game/objects/traps/dart.c
@@ -3,7 +3,7 @@
 #include "game/effects.h"
 #include "game/effects/blood.h"
 #include "game/items.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "game/random.h"
 #include "game/room.h"
 #include "game/sound.h"
