+++ src/game/objects/traps/flame.c
@@ -4,7 +4,7 @@
 #include "game/effects.h"
 #include "game/items.h"
 #include "game/lara.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "game/room.h"
 #include "game/sound.h"
 #include "global/vars.h"
