+++ src/game/objects/traps/lava.c
@@ -3,7 +3,7 @@
 #include "game/effects.h"
 #include "game/items.h"
 #include "game/lara.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "game/random.h"
 #include "game/room.h"
 #include "game/sound.h"
