+++ src/game/objects/traps/midas_touch.c
@@ -3,7 +3,7 @@
 #include "game/input.h"
 #include "game/inv.h"
 #include "game/lara/lara.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "global/const.h"
 #include "global/vars.h"
