+++ src/game/objects/traps/rolling_ball.c
@@ -5,7 +5,7 @@
 #include "game/gamebuf.h"
 #include "game/items.h"
 #include "game/lara/lara.h"
-#include "game/objects.h"
+#include "game/objects/common.h"
 #include "game/random.h"
 #include "game/room.h"
 #include "global/const.h"
