+++ src/game/room.c
@@ -4,9 +4,9 @@
 #include "game/items.h"
 #include "game/lot.h"
 #include "game/music.h"
-#include "game/objects/keyhole.h"
-#include "game/objects/pickup.h"
-#include "game/objects/switch.h"
+#include "game/objects/general/keyhole.h"
+#include "game/objects/general/pickup.h"
+#include "game/objects/general/switch.h"
 #include "game/objects/traps/lava.h"
 #include "game/shell.h"
 #include "game/sound.h"
