+++ src/game/savegame.c
@@ -6,8 +6,8 @@
 #include "game/inv.h"
 #include "game/items.h"
 #include "game/objects/creatures/pod.h"
-#include "game/objects/pickup.h"
-#include "game/objects/puzzle_hole.h"
+#include "game/objects/general/pickup.h"
+#include "game/objects/general/puzzle_hole.h"
 #include "game/objects/traps/movable_block.h"
 #include "game/objects/traps/rolling_block.h"
 #include "game/room.h"
