+++ src/game/setup.c
@@ -6,11 +6,7 @@
 #include "game/inv.h"
 #include "game/lara.h"
 #include "game/level.h"
-#include "game/objects.h"
-#include "game/objects/boat.h"
-#include "game/objects/bridge.h"
-#include "game/objects/cabin.h"
-#include "game/objects/cog.h"
+#include "game/objects/common.h"
 #include "game/objects/creatures/ape.h"
 #include "game/objects/creatures/bacon_lara.h"
 #include "game/objects/creatures/baldy.h"
@@ -34,7 +30,6 @@
 #include "game/objects/creatures/torso.h"
 #include "game/objects/creatures/trex.h"
 #include "game/objects/creatures/wolf.h"
-#include "game/objects/door.h"
 #include "game/objects/effects/blood.h"
 #include "game/objects/effects/body_part.h"
 #include "game/objects/effects/bubble.h"
@@ -46,14 +41,19 @@
 #include "game/objects/effects/splash.h"
 #include "game/objects/effects/twinkle.h"
 #include "game/objects/effects/waterfall.h"
-#include "game/objects/keyhole.h"
-#include "game/objects/misc.h"
-#include "game/objects/pickup.h"
-#include "game/objects/puzzle_hole.h"
-#include "game/objects/save_crystal.h"
-#include "game/objects/scion.h"
-#include "game/objects/switch.h"
-#include "game/objects/trapdoor.h"
+#include "game/objects/general/boat.h"
+#include "game/objects/general/bridge.h"
+#include "game/objects/general/cabin.h"
+#include "game/objects/general/cog.h"
+#include "game/objects/general/door.h"
+#include "game/objects/general/keyhole.h"
+#include "game/objects/general/misc.h"
+#include "game/objects/general/pickup.h"
+#include "game/objects/general/puzzle_hole.h"
+#include "game/objects/general/save_crystal.h"
+#include "game/objects/general/scion.h"
+#include "game/objects/general/switch.h"
+#include "game/objects/general/trapdoor.h"
 #include "game/objects/traps/damocles_sword.h"
 #include "game/objects/traps/dart.h"
 #include "game/objects/traps/falling_block.h"
